Move DatabaseFeatures to features module
