+++ django_ibmi/features.py
@@ -1,0 +1,42 @@
+from django.db.backends.base.features import BaseDatabaseFeatures
+
+
+class DatabaseFeatures(BaseDatabaseFeatures):
+    can_use_chunked_reads = True
+
+    # Save point is supported by DB2.
+    uses_savepoints = True
+
+    # Custom query class has been implemented
+    # django.db.backends.db2.query.query_class.DB2QueryClass
+    uses_custom_query_class = True
+
+    # transaction is supported by DB2
+    supports_transactions = True
+
+    supports_tablespaces = True
+
+    uppercases_column_names = True
+    interprets_empty_strings_as_nulls = False
+    allows_primary_key_0 = True
+    can_defer_constraint_checks = False
+    supports_forward_references = False
+    requires_rollback_on_dirty_transaction = True
+    supports_regex_backreferencing = True
+    supports_timezones = False
+    has_bulk_insert = False
+    has_select_for_update = True
+    supports_long_model_names = False
+    can_distinct_on_fields = False
+    supports_paramstyle_pyformat = False
+    supports_sequence_reset = True
+    # DB2 doesn't take default values as parameter
+    requires_literal_defaults = True
+    has_case_insensitive_like = True
+    can_introspect_big_integer_field = True
+    can_introspect_boolean_field = False
+    can_introspect_positive_integer_field = False
+    can_introspect_small_integer_field = True
+    can_introspect_null = True
+    can_introspect_ip_address_field = False
+    can_introspect_time_field = True
